@@ -1,6 +1,6 @@
 module github.com/vdbulcke/oidc-client-demo
 
-go 1.18
+go 1.20
 
 require (
 	github.com/coreos/go-oidc/v3 v3.4.0
