@@ -2,7 +2,7 @@ package oidcclient
 
 import (
 	"fmt"
-	"io/ioutil"
+	"io"
 	"os"
 
 	"github.com/go-playground/validator"
@@ -116,7 +116,7 @@ func ParseConfig(configFile string) (*OIDCClientConfig, error) {
 	}
 
 	defer file.Close()
-	data, err := ioutil.ReadAll(file)
+	data, err := io.ReadAll(file)
 	if err != nil {
 		return nil, err
 	}
